move identity to its own crate
