+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "ipld/resolver", "testing/e2e"]
+members = [".", "ipld/resolver", "testing/e2e", "identity"]
 
 [workspace.package]
 authors = ["Protocol Labs"]
@@ -24,7 +24,7 @@ indoc = "2.0.0"
 log = { workspace = true }
 reqwest = { version = "0.11.13", features = ["json"] }
 serde = { workspace = true }
-serde_json = { version = "1.0.91", features = ["raw_value"] }
+serde_json = { workspace = true }
 cid = { version = "0.8.3", default-features = false, features = ["serde-codec"] }
 tokio = { workspace = true }
 tokio-graceful-shutdown = "0.12.1"
@@ -33,7 +33,7 @@ derive_builder = "0.12.0"
 num-traits = "0.2.15"
 num-derive = "0.3.3"
 env_logger = "0.10.0"
-base64 = "0.21.0"
+base64 = { workspace = true }
 strum = { version = "0.24", features = ["derive"] }
 toml = "0.7.2"
 url = { version = "2.3.1", features = ["serde"] }
@@ -41,19 +41,8 @@ warp = "0.3.3"
 bytes = "1.4.0"
 serde_bytes = "0.11.9"
 clap = { version = "4.1.4", features = ["env", "derive"] }
-thiserror = "1.0.38"
+thiserror = { workspace = true }
 serde_tuple = "0.5.0"
-blake2b_simd = { workspace = true }
-rand = { workspace = true }
-ahash = "0.8"
-libsecp256k1 = "0.7"
-argon2 = "0.5"
-xsalsa20poly1305 = "0.9"
-serde_ipld_dagcbor = "0.2"
-libc = "0.2"
-quickcheck = { workspace = true }
-quickcheck_macros = { workspace = true }
-bls-signatures = { version = "0.12", default-features = false, features = ["blst"] }
 
 fvm_shared = { workspace = true }
 fil_actors_runtime = { workspace = true }
@@ -63,22 +52,26 @@ ipc-gateway = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
 primitives = { workspace = true }
 
+ipc-identity = { path = "identity/." }
+
 [dev-dependencies]
 tempfile = "3.4.0"
 
 [workspace.dependencies]
 anyhow = "1.0"
+base64 = "0.21.0"
 lazy_static = "1.4"
 log = "0.4"
 env_logger = "0.10"
 prometheus = "0.13"
 serde = { version = "1.0", features = ["derive"] }
-thiserror = "1.0"
 tokio = { version = "1.16", features = ["full"] }
+thiserror = "1.0.38"
 quickcheck = "1"
 quickcheck_macros = "1"
 blake2b_simd = "1.0"
 rand = "0.8"
+serde_json = { version = "1.0.91", features = ["raw_value"] }
 
 fvm_ipld_blockstore = "0.1"
 fvm_ipld_encoding = "0.3"
