+++ identity/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "ipc-identity"
+description = "IPC key management and identity"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+license-file.workspace = true
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = { workspace = true }
+base64 = { workspace = true }
+blake2b_simd = { workspace = true }
+rand = { workspace = true }
+ahash = "0.8"
+libsecp256k1 = "0.7"
+argon2 = "0.5"
+xsalsa20poly1305 = "0.9"
+serde_ipld_dagcbor = "0.2"
+libc = "0.2"
+quickcheck = { workspace = true }
+quickcheck_macros = { workspace = true }
+bls-signatures = { version = "0.12", default-features = false, features = ["blst"] }
+serde_json = { workspace = true }
+serde = { workspace = true }
+log = { workspace = true }
+thiserror = { workspace = true }
+
+fvm_shared = { workspace = true }
+
+ipc-sdk = { workspace = true }
+
+
+[dev-dependencies]
+tokio = { workspace = true }
+log = { workspace = true }
+tempfile = "3.4.0"
