+++ identity/src/wallet_helpers.rs
