+++ identity/tests/keystore_encrypted_old/keystore
